Day 11, part 2

+++ 2023/Day11.h
@@ -7,7 +7,7 @@ private:
 	const std::string testFile2 = "inputs/day11_part2_test.txt";
 	const std::string puzzleFile = "inputs/day11.txt";
 	int calculatePuzzle1(std::vector<std::string> input);
-	int calculatePuzzle2(std::vector<std::string> input);
+	int64_t calculatePuzzle2(std::vector<std::string> input);
 public:
 	void puzzle1();
 	void puzzle2();
